+++ src/ObjWriting/Game/IW3/AssetDumpers/AssetDumperXModel.cpp
@@ -2,10 +2,10 @@
 
 #include "Game/IW3/CommonIW3.h"
 #include "Math/Quaternion.h"
-#include "Model/XModel/XModelExportWriter.h"
 #include "ObjWriting.h"
 #include "Utils/HalfFloat.h"
 #include "Utils/QuatInt16.h"
+#include "XModel/Export/XModelExportWriter.h"
 
 #include <cassert>
 #include <sstream>
