+++ src/ObjWriting/Game/IW5/AssetDumpers/AssetDumperXModel.cpp
@@ -2,10 +2,10 @@
 
 #include "Game/IW5/CommonIW5.h"
 #include "Math/Quaternion.h"
-#include "Model/XModel/XModelExportWriter.h"
 #include "ObjWriting.h"
 #include "Utils/HalfFloat.h"
 #include "Utils/QuatInt16.h"
+#include "XModel/Export/XModelExportWriter.h"
 
 #include <cassert>
