+++ src/ObjWriting/Game/T6/AssetDumpers/AssetDumperXModel.cpp
@@ -2,10 +2,10 @@
 
 #include "Game/T6/CommonT6.h"
 #include "Math/Quaternion.h"
-#include "Model/XModel/XModelExportWriter.h"
 #include "ObjWriting.h"
 #include "Utils/HalfFloat.h"
 #include "Utils/QuatInt16.h"
+#include "XModel/Export/XModelExportWriter.h"
 
 #include <cassert>
 #include <sstream>
