+++ src/ObjWriting/XModel/Export/XModelExportWriter.cpp
